==> 00_Formatting_count_matrix.R <==
#Script generated by: Aaron D. Ramirez-Sanchez 
#Goal: To format RNA count data with new names and filter out outliers

##ENVIRONMENT
library(Matrix)

##FUNCTIONS

##INPUT

#Count matrix
EL.rna.matrix <- as.matrix(read.csv(file = "./data/raw/gene_expression/epithelialOslo_data2.expression.genelevel.v75.htseq.txt.table", sep="\t", row.names = "probe"))

#Sample new names 
sample.names <- read.delim(file = "./data/processed/metadata/sample_names.tsv", header = T)

##MAIN
#sorting matrix
EL.rna.matrix <- EL.rna.matrix[,as.character(sample.names$sample)]
colnames(EL.rna.matrix) <- sample.names$sample.new

##OUTPUT
write.table(EL.rna.matrix, file = "./data/processed/gene_expression/RNA_matrix.table", 
            sep = "\t", 
            col.names = NA, 
            row.names = T, 
            quote = F)
