==> 00_Formatting_metadata.R <==
#Script generated by: Aaron D. Ramirez-Sanchez 
#Goal: To format metadata and create a sample_sample conversion

##ENVIRONMENT
library(rjson)
library(dplyr)

##FUNCTIONS

##INPUT

#Metadata
coldata <- read.csv(file = "./data/raw/metadata/20230220_Metadata_OsloRNA.csv")
#GC content
result.json <- fromJSON(file = "./data/raw/metadata/GC_content.json", simplify = F)
#Samplesheet
sample.sheet <- read.csv(file = "./data/raw/metadata/202004-GS-SampleCode-SeqAllv4.csv", sep = ",")

##MAIN

#Adding extracolumn for sex based on chrY
coldata$sex2 <- as.factor(ifelse(coldata$Percentage.Y.Chr < 0.03, paste("F"), paste("M")))
#Adding age to missing samples
coldata$age2 <- as.numeric(coldata$age)
coldata[is.na(coldata$age2),]$age2 <- floor(runif(3, 19, 86))

#Adding column with sex and condition
coldata$condition.sex <- paste(coldata$condition, coldata$sex2, sep = ".")
coldata$condition.sex <- as.factor(coldata$condition.sex)

#Adding a column of GC
#processing json file
result.json.df <- as.data.frame(result.json)
GC.content <- result.json$datasets[[10]]
names(GC.content) <- result.json$samples[[10]]
GC.content <- unlist(GC.content)
#processing sample sheet
sample.sheet$long.code <- gsub("-",".", sample.sheet$long.code)
samples.GC <- data.frame(sample_long = names(GC.content), GC = as.integer(GC.content))
samples.GC$long.code <- gsub("-", ".", samples.GC$sample_long)
samples.GC$long.code <- gsub(".*_L._", "", samples.GC$long.code)
samples.GC$long.code <- gsub(".{2}$", "", samples.GC$long.code)
samples.GC$long.code <- gsub(" ", "", samples.GC$long.code)
df.GC <- left_join(sample.sheet, samples.GC, by = "long.code")
df.GC <- aggregate(.~sample_name, FUN=mean, data=df.GC[,c(4,7)])
#adding GC to metadata
coldata$condition <- as.factor(coldata$condition)
filtered.GC <- df.GC[df.GC$sample_name %in% as.character(coldata$sample), ]
coldata$GC <- filtered.GC[match(as.character(coldata$sample), filtered.GC$sample_name), ]$GC

#Creating table with sample name for reference
sample.names <- coldata[,c("sample", "sample2", "sample.new")]

##OUTPUT

write.table(coldata, file = "./data/processed/metadata/metadata.tsv", 
            sep = "\t", 
            col.names = T, 
            row.names = F, 
            quote = F)
write.table(sample.names, file = "./data/processed/metadata/sample_names.tsv", 
            sep = "\t", 
            col.names = T, 
            row.names = F, 
            quote = F)
