==> 04_ZeroVar_count_matrix.R <==
#Script generated by: Aaron D. Ramirez-Sanchez 
#Goal: To format metadata and create a sample_sample conversion

##ENVIRONMENT
library(Matrix)
library(data.table)
library(dplyr)
library(DESeq2)
library(Rfast)
library(edgeR)

##FUNCTIONS

##INPUT

#Reading dds object
dds <- readRDS("./output/objects/dds.RDS")

##MAIN

#removing probes with zero variance
RNA.matrix <- counts(dds)
RNA.matrix <- RNA.matrix[!rowVars(RNA.matrix) == 0,]

#Generating normalized matrix without variance probes
D <- DGEList(counts=RNA.matrix)
d <- calcNormFactors(D)
scalar <- d$samples$lib.size*d$samples$norm.factors/exp(mean(log(d$samples$lib.size*d$samples$norm.factors)))
scal.mat <- outer(rep(1,nrow(d$counts)), scalar)
scaled.counts <- d$counts/scal.mat

cpm.D <- cpm(D)

##OUTPUT

write.table(scaled.counts, file = "./output/tables/gene_expression/RNA_matrix_zeroVar.table", sep = "\t", row.names=TRUE, quote=FALSE, col.names = NA)
