==> 05_Heatmaps_and_clusters.R <==
#Script generated by: Aaron D. Ramirez-Sanchez 
#Goal: To format metadata and create a sample_sample conversion

##ENVIRONMENT
library(ComplexHeatmap)
library(viridis)
library(circlize)
library(dplyr)
library(factoextra)

##FUNCTIONS
theme.plain <- function(p, base_size = 11, base_family = "ArialMT") {
  p <- p + theme_linedraw(base_size = base_size, base_family = base_family) %+replace%
    theme(panel.background = element_blank(),
          #panel.border = element_blank(),
          panel.grid.major = element_blank(),
          panel.grid.minor = element_blank(),
          axis.line = element_line(color="black", size=0.75),
          axis.ticks = element_line(size=0.75),
          axis.text = element_text(size=base_size, family="ArialMT", face="plain"),
          strip.background = element_blank(),
          strip.text = element_text(size=base_size, family="ArialMT", face="plain"),
          legend.key = element_blank(),
          legend.text = element_text(size=base_size, family = "ArialMT", face="plain"),
          complete = TRUE,
          plot.title = element_text(hjust=0.5))
  return(p)
}
##INPUT

#Loading RNA matrix
RNA.DEG <- as.matrix(read.csv(file = "./output/tables/gene_expression/RNA_matrix_zeroVar.table", sep="\t", row.names = 1))

#Loading metadata
annotation.data <- read.delim(file = "./data/processed/metadata/metadata.tsv", sep = "\t", row.names = "sample.new")
annotation.data$marsh.2 <- as.factor(annotation.data$marsh.2)

#Loading merged DEG 
DEG.merged <- read.csv(file = "./output/tables/DEG/comparisons_merged.csv", header = T, stringsAsFactors = F)

#Loading cell proportions
#cell.proportions <- read.table('./output/tables/FACS/cell_proportions_scaled.2groups.txt')

#Filters used for the DEG
L2FC.filter <- 1.0
p.adj.filter <- 0.01

#Colors
col_marsh <- viridis(4)

##MAIN
#Filtering data
DEG.merged <- DEG.merged[(abs(DEG.merged$log2FoldChange) > L2FC.filter & !is.na(DEG.merged$log2FoldChange)) & 
                           (DEG.merged$padj < p.adj.filter & !is.na(DEG.merged$padj)), ]
DEG.list <- unique(DEG.merged$ENSEMBL)

RNA.DEG <- RNA.DEG[match(DEG.list, rownames(RNA.DEG)),]

#Scaling matrix
base_mean <-  rowMeans(RNA.DEG)
mat_scaled <-  t(scale(t(RNA.DEG)))

#Checking number of k-means
k.gap.genes <- fviz_nbclust(mat_scaled[DEG.list,], kmeans, 
                            method = "gap_stat", 
                            nboot = 100)



k.silhouette.genes <- fviz_nbclust(mat_scaled[DEG.list,], kmeans, 
                                   method = "silhouette", 
                                   nboot = 100)
k.wss.genes <- fviz_nbclust(mat_scaled[DEG.list,], kmeans, 
                            method = "wss", 
                            nboot = 100)
pdf(file = "./output/plots/Optimal_k_clusters_genes_gap.pdf", height = 4, width = 4)
theme.plain(k.gap.genes)
theme.plain(k.silhouette.genes)
theme.plain(k.wss.genes)
dev.off()

k.gap.samples <- fviz_nbclust(t(mat_scaled[DEG.list,]), kmeans, 
                              method = "gap_stat", 
                              nboot = 100)
k.silhouette.samples <- fviz_nbclust(t(mat_scaled[DEG.list,]), kmeans, 
                                     method = "silhouette", 
                                     nboot = 100)
k.wss.samples <- fviz_nbclust(t(mat_scaled[DEG.list,]), kmeans, 
                              method = "wss", 
                              nboot = 100)
pdf(file = "./output/plots/Optimal_k_clusters_samples_gap.pdf", height = 4, width = 4)
theme.plain(k.gap.samples)
theme.plain(k.silhouette.samples)
theme.plain(k.wss.samples)
dev.off()

#type <-  gsub("s\\d+_", "", colnames(RNA.DEG))
#ha <-  HeatmapAnnotation(df = data.frame(type = type))
names_marsh <- sort(unique(annotation.data[,"marsh.2"]))
names(col_marsh) <- names_marsh

column.ha <- HeatmapAnnotation(df = annotation.data[, c("condition", "marsh.2")],
                               col = list(condition = c("UCD" = "#D9514EFF", "CTRL" = "#2A2B2DFF", "TCD" = "#2DA8D8FF"),
                                          marsh.2 = col_marsh))

m <- Heatmap(mat_scaled , 
             top_annotation = column.ha,
             name = "Z-score", 
             km = 4, 
             column_km = 3,
             col = colorRamp2(c(-2, 0, 2), c("#2166ac", "white", "#b2182b")), 
             show_row_names = FALSE, 
             show_column_names = TRUE,
             use_raster = F,
             column_names_gp = gpar(fontsize = 6))
#Since clustering use pseudorandom numbers, setting sed to get reproducible results and using draw to get consistent results
set.seed(1234)
ht_list = m
ht_list = draw(ht_list)

#Saving groups for rows and columns
clusterlist <- row_order(ht_list)
columnlist <- column_order(ht_list)

#Manually adjusting heatmap

#First adjusting rows, having first groups genes very strong for CeD [4] > moderate [3] > small group downregulated [1] > big group downregulated [2] in CeD
clu_df <- lapply(names(clusterlist), function(i){
  out <- data.frame(GeneID = rownames(mat_scaled)[clusterlist[[i]]],
                    Cluster = paste0("cluster", i),
                    stringsAsFactors = FALSE)
  return(out)
}) %>%  
  do.call(rbind, .)
clu_df$Cluster.original<- clu_df$Cluster
#clu_df$Cluster.original<- clu_df$Cluster
clu_df_reordered <- gsub("cluster", "Cluster_", clu_df[match(rownames(mat_scaled), clu_df$GeneID),]$Cluster)
clu_df_reordered <- gsub("Cluster_1", "Cluster_c", clu_df_reordered)
clu_df_reordered <- gsub("Cluster_2", "Cluster_d", clu_df_reordered)
clu_df_reordered <- gsub("Cluster_3", "Cluster_a", clu_df_reordered)
clu_df_reordered <- gsub("Cluster_4", "Cluster_b", clu_df_reordered)

clu_df_reordered <- gsub("Cluster_a", "Cluster_1", clu_df_reordered)
clu_df_reordered <- gsub("Cluster_b", "Cluster_2", clu_df_reordered)
clu_df_reordered <- gsub("Cluster_c", "Cluster_3", clu_df_reordered)
clu_df_reordered <- gsub("Cluster_d", "Cluster_4", clu_df_reordered)

clu_df_reordered <- factor(clu_df_reordered, levels = c("Cluster_1", "Cluster_2", "Cluster_3", "Cluster_4"))
clu_df <- clu_df[match(rownames(mat_scaled), clu_df$GeneID),]
clu_df$Cluster <- clu_df_reordered

#Second adjusting cols, having first group of controls, moderates and CeD
col_df <- lapply(names(columnlist), function(i){
  out <- data.frame(GeneID = colnames(mat_scaled)[columnlist[[i]]],
                    Cluster = paste0("cluster", i),
                    stringsAsFactors = FALSE)
  return(out)
}) %>%  
  do.call(rbind, .)
col_df$Cluster.original<- col_df$Cluster
col_df$Cluster <- gsub("cluster1", "clusterX", col_df$Cluster)
col_df$Cluster <- gsub("cluster2", "cluster1", col_df$Cluster)
col_df$Cluster <- gsub("clusterX", "cluster2", col_df$Cluster)
col_df_reordered <- gsub("cluster", "Group_", col_df[match(colnames(mat_scaled), col_df$GeneID),]$Cluster)

col_df <- col_df[match(colnames(mat_scaled), col_df$GeneID),]
col_df$Cluster <- col_df_reordered

#Adding extra metadata for the heatmaps
#adding metadata for APOA4 and KI67
extra.metadata <- read.csv(file = "./data/processed/metadata/metaData_cryptRatios_131022.txt", sep = "\t", header = T, stringsAsFactors = F)
extra.metadata$sample <- gsub("_", "", extra.metadata$sample)
rownames(extra.metadata) <- extra.metadata$sample
extra.metadata <- extra.metadata[annotation.data$sample, ]
rownames(extra.metadata) <- rownames(annotation.data)

#creating matrix for plots
mx = as.matrix(extra.metadata[,15:16])
mx = t(apply(mx, 1, function(x) x/sum(x)))
#creating colors for crypts ratio
col_fun <-  colorRamp2(c(1, 1.5, 2), c("#A1E1FF", "#52B3EB", "#005A99"))
colors <- col_fun(extra.metadata[,17])

column.ha.2 = HeatmapAnnotation(df = annotation.data[, c("condition", "marsh.2")],
                                APOA4_KI67 = anno_barplot(as.matrix(extra.metadata[,17]), 
                                                          height = unit(2, "cm"), 
                                                          gp = gpar(fill = col_fun(extra.metadata[,17])),
                                                          border = F),
                                col = list(condition = c("UCD" = "#D9514EFF", 
                                                         "CTRL" = "#2A2B2DFF", 
                                                         "TCD" = "#2DA8D8FF"),
                                           marsh.2 = col_marsh),
                                annotation_legend_param = list(
                                  condition = list(
                                    title = "Condition"
                                  ),
                                  marsh.2 = list(
                                    title = "Marsh"
                                  )
                                  )
)

#Plotting custom heatmap with extra features
m2 <- Heatmap(mat_scaled , 
              top_annotation = column.ha.2,
              name = "Z-score", 
              #km = 4, 
              #column_km = 3,
              split = clu_df_reordered,
              cluster_row_slices = FALSE,
              column_split = col_df_reordered,
              col = colorRamp2(c(-2, 0, 2), c("#2166ac", "white", "#b2182b")), 
              show_row_names = FALSE, 
              show_column_names = TRUE,
              use_raster = F,
              column_names_gp = gpar(fontsize = 6))
ht_list2 = m2
ht_list2 = draw(ht_list2)

#Creating extended metadata with new columns for crypts ration and groups of patients
extended_metadata <- annotation.data
extended_metadata <- cbind(extended_metadata)
extended_metadata <- cbind(extended_metadata, extra.metadata[,15:17], col_df)

#Creating extended merged DEG
extended_DEG.merged <- DEG.merged
clu_df.tmp <- clu_df
clu_df.tmp$ENSEMBL <- clu_df.tmp$GeneID

extended_DEG.merged <- merge(extended_DEG.merged, clu_df.tmp, by = "ENSEMBL")

##OUTPUT

#Saving Heatmap
pdf(file = "./output/plots/heatmap_DEG.pdf", width=12, height=8,  family = "ArialMT")
draw(ht_list2)
dev.off()

#Saving clusters of genes and extended metadata
write.table(clu_df, 
          file="./data/processed/metadata/clusters_DEG.tsv", 
          row.names = T, col.names = NA, quote = F, sep = "\t")

write.table(extended_metadata, 
          file="./data/processed/metadata/extended_metadata.tsv", 
          row.names = T, col.names = NA, quote = F, sep = "\t")

write.csv(extended_DEG.merged, 
          file="./output/tables/DEG/extended_comparisons_merged.csv", 
          row.names = F, quote = F)
