==> 08_PCA_plots.R <==
#Script generated by: Aaron D. Ramirez-Sanchez 
#Goal: To format metadata and create a sample_sample conversion

##ENVIRONMENT
library(DESeq2)
library(ggplot2)
library(ggpubr)
library(PCAtools)
library(viridis)
library(ggrepel)
library(GGally) 
library(RColorBrewer)

##FUNCTIONS
#FUN: plotPCA3.4, gets PCA1-4
plotPCA3.4 <- function (object, intgroup = "condition", ntop = 500, returnData = FALSE) 
{
  rv <- rowVars(assay(object))
  select <- order(rv, decreasing = TRUE)[seq_len(min(ntop, 
                                                     length(rv)))]
  pca <- prcomp(t(assay(object)[select, ]))
  percentVar <- pca$sdev^2/sum(pca$sdev^2)
  if (!all(intgroup %in% names(colData(object)))) {
    stop("the argument 'intgroup' should specify columns of colData(dds)")
  }
  intgroup.df <- as.data.frame(colData(object)[, intgroup, 
                                               drop = FALSE])
  group <- if (length(intgroup) > 1) {
    factor(apply(intgroup.df, 1, paste, collapse = " : "))
  }
  else {
    colData(object)[[intgroup]]
  }
  d <- data.frame(PC1 = pca$x[, 1], PC2 = pca$x[, 2], PC3 = pca$x[, 3], PC4 = pca$x[, 4], group = group, 
                  intgroup.df, name = colnames(object))
  if (returnData) {
    attr(d, "percentVar") <- percentVar[1:4]
    return(d)
  }
  ggplot(data = d, aes_string(x = "PC3", y = "PC4", color = "group")) + 
    geom_point(size = 3) + xlab(paste0("PC3: ", round(percentVar[1] * 
                                                        100), "% variance")) + ylab(paste0("PC4: ", round(percentVar[2] * 
                                                                                                            100), "% variance")) + coord_fixed()
}

#FUN: pairedPCA.plots.pdf, function to make plot for categorical variables
pairedPCA.plots.pdf <- function(df.pca, test) {
  #Making correlation plots
  p <- ggpairs(df.pca[,c(1:4)],
               aes(color = df.pca[[test]], alpha = 0.33),
               title = test)
  return(theme.plain(p))
}

#FUN, theme.plain
theme.plain <- function(p, base_size = 11, base_family = "ArialMT") {
  p <- p + theme_linedraw(base_size = base_size, base_family = base_family) %+replace%
    theme(panel.background = element_blank(),
          #panel.border = element_blank(),
          panel.grid.major = element_blank(),
          panel.grid.minor = element_blank(),
          axis.line = element_line(color="black", size=0.75),
          axis.ticks = element_line(size=0.75),
          axis.text = element_text(size=base_size, family="ArialMT", face="plain"),
          strip.background = element_blank(),
          strip.text = element_text(size=base_size, family="ArialMT", face="plain"),
          legend.key = element_blank(),
          legend.text = element_text(size=base_size, family = "ArialMT", face="plain"),
          complete = TRUE,
          plot.title = element_text(hjust=0.5))
  return(p)
}

##INPUT

dds <- readRDS("./output/objects/dds.RDS")

#colors: Danger Red, Tap Shoe and Blue Blossom
pallete_condition <- c("#D9514EFF", "#2A2B2DFF", "#2DA8D8FF")

##MAIN
#transforming data
vsd <- vst(dds, blind=FALSE)

sampleDists <- dist(t(assay(vsd)))

pcaData3.4 <- plotPCA3.4(vsd, intgroup= names(colData(dds)), returnData=TRUE)
pcaData3.4$marsh.2 <- as.character(pcaData3.4$marsh.2)
percentVar3.4 <- round(100 * attr(pcaData3.4, "percentVar"))


##OUTPUT
#savinf pdf
pdf(file = "./output/plots/PCA_plots.pdf", width=8, height=8,  family = "ArialMT")

pairedPCA.plots.pdf(pcaData3.4, "condition") + 
  scale_color_manual(values = pallete_condition[c(2,3,1)]) + 
  scale_fill_manual(values = pallete_condition[c(2,3,1)])

pairedPCA.plots.pdf(pcaData3.4, "dtt") + 
  scale_color_manual(values = c("#00A4CCFF","#F95700FF")) + 
  scale_fill_manual(values = c("#00A4CCFF","#F95700FF"))

pairedPCA.plots.pdf(pcaData3.4, "PPI") + 
  scale_color_manual(values = c("#00A4CCFF","#F95700FF")) + 
  scale_fill_manual(values = c("#00A4CCFF","#F95700FF"))

pairedPCA.plots.pdf(pcaData3.4, "H.pylori") + 
  scale_color_manual(values = c("#00A4CCFF","#F95700FF")) + 
  scale_fill_manual(values = c("#00A4CCFF","#F95700FF"))

pairedPCA.plots.pdf(pcaData3.4, "sex") + 
  scale_color_manual(values = c("#00A4CCFF","#F95700FF")) + 
  scale_fill_manual(values = c("#00A4CCFF","#F95700FF"))

p <- ggpairs(pcaData3.4[,c(1:4)],
             aes(color = pcaData3.4$Percentage.Y.Chr),
             upper = list(continuous = "blank", combo = "box_no_facet", discrete = "count", na =
                            "na"),
             title = "Percentage.Y.Chr",
             legend = c(2,1))
p <- p + scale_color_viridis(discrete = F, direction = -1)
theme.plain(p) +
  theme(legend.position = "bottom")

p <- ggpairs(pcaData3.4[,c(1:4)],
             aes(color = pcaData3.4$marsh, alpha = 0.33),
             upper = list(continuous = "blank", combo = "box_no_facet", discrete = "count", na = "na"),
             title = "marsh",
             legend = c(1,1))
p <- p +  scale_color_viridis(discrete = T, direction = -1, alpha = 0.5) +
  scale_fill_viridis(discrete = T, direction = -1, alpha = 0.5)
theme.plain(p) +
  theme(legend.position = "bottom")

pairedPCA.plots.pdf(pcaData3.4, "sequencing.batch")

p <- ggpairs(pcaData3.4[,c(1:4)],
             aes(color = pcaData3.4$marsh.2, alpha = 0.33),
             upper = list(continuous = "blank", combo = "box_no_facet", discrete = "count", na = "na"),
             title = "marsh.2",
             legend = c(1,1))
p <- p +  scale_color_viridis(discrete = T, direction = -1, alpha = 0.5) +
  scale_fill_viridis(discrete = T, direction = -1, alpha = 0.5)
theme.plain(p) +
  theme(legend.position = "bottom")

pairedPCA.plots.pdf(pcaData3.4, "sequencing.batch")

p <- ggpairs(pcaData3.4[,c(1:4)],
             aes(color = as.factor(pcaData3.4$batch.RNA.isolation), alpha = 0.33),
             upper = list(continuous = "blank", combo = "box_no_facet", discrete = "count", na =
                            "na"),
             title = "batch.RNA.isolation",
             legend = c(1,1))
theme.plain(p) + theme(legend.position = "bottom")

p <- ggpairs(pcaData3.4[,c(1:4)],
             aes(color = pcaData3.4$age),
             upper = list(continuous = "blank", combo = "box_no_facet", discrete = "count", na =
                            "na"),
             title = "age",
             legend = c(2,1))
p <- p + scale_color_viridis(discrete = F, direction = -1)
theme.plain(p) +
  theme(legend.position = "bottom")

p <- ggpairs(pcaData3.4[,c(1:4)],
             aes(color = log10(pcaData3.4$concentration.RNA.GS)),
             upper = list(continuous = "blank", combo = "box_no_facet", discrete = "count", na =
                            "na"),
             title = "concentration.RNA.GS",
             legend = c(2,1))
p <- p + scale_color_viridis(discrete = F, direction = -1)
theme.plain(p) +
  theme(legend.position = "bottom")

p <- ggpairs(pcaData3.4[,c(1:4)],
             aes(color = pcaData3.4$Unmapped.Reads.P),
             upper = list(continuous = "blank", combo = "box_no_facet", discrete = "count", na =
                            "na"),
             title = "Unmapped.Reads.P",
             legend = c(2,1))
p <- p + scale_color_viridis(discrete = F, direction = -1)
theme.plain(p) +
  theme(legend.position = "bottom")

p <- ggpairs(pcaData3.4[,c(1:4)],
             aes(color = pcaData3.4$Total.reads),
             upper = list(continuous = "blank", combo = "box_no_facet", discrete = "count", na =
                            "na"),
             title = "Total.reads",
             legend = c(2,1))
p <- p + scale_color_viridis(discrete = F, direction = -1)
theme.plain(p) +
  theme(legend.position = "bottom")

dev.off()

pdf(file = "./output/plots/PCA_plots_sex.pdf", width=6, height=5,  family = "ArialMT")
p <- ggplot(pcaData3.4, aes(PC2, PC3, color=sex)) +
  geom_point(size=3) +
  xlab(paste0("PC2: ",percentVar3.4[2],"% variance")) +
  ylab(paste0("PC3: ",percentVar3.4[3],"% variance")) + 
  geom_text_repel(aes(label= as.character(sample))) +
  coord_fixed()
theme.plain(p) + coord_fixed(1/1.2)
dev.off()


pdf(file = "./output/plots/PCA_plots_marsh.pdf", width=6, height=5,  family = "ArialMT")
p <- ggplot(pcaData3.4, aes(PC1, PC2, color=marsh.2)) +
  geom_point(size=3) +
  xlab(paste0("PC1: ",percentVar3.4[1],"% variance")) +
  ylab(paste0("PC2: ",percentVar3.4[2],"% variance")) + 
  geom_text_repel(aes(label= as.character(sample))) +
  coord_fixed()
p <- p +  scale_color_viridis(discrete = T, direction = -1, alpha = 0.5) +
  scale_fill_viridis(discrete = T, direction = -1, alpha = 0.5)
theme.plain(p)
dev.off()
