==> 10_Enrichment_analysis_clusters_DEG.R <==
#Script generated by: Aaron D. Ramirez-Sanchez 
#Goal: To perform enrichment analysis tables and objects

##ENVIRONMENT
library(clusterProfiler)
library(ReactomePA)
library(enrichplot)
library(msigdbr)


##FUNCTIONS
# Function to put in compareCluster
gse_fun <- function(gene_id = ENTREZID, term = msigdbr_t2g) {
  enricher(gene = gene_id, TERM2GENE = term, pvalueCutoff = 0.05)
}

##INPUT

#reading merged DEG 
DEG.merged <- read.csv(file = "./output/tables/DEG/clusters_comparisons_merged.csv", header = T, stringsAsFactors = F)
# Getting human hallmark dataset
h_gene_set <- msigdbr(species = "Homo sapiens", category = "H")
msigdbr_t2g <- h_gene_set %>% 
  dplyr::distinct(gs_name, entrez_gene) %>% 
  as.data.frame()
#Filters used for the DEG
L2FC.filter <- 1.0
p.adj.filter <- 0.01

RNA.ref <- read.table("/Users/R_projects/Oslo/results/eQTL/references/Homo_sapiens.GRCh37.75.genes.bed", sep = "\t", header = F)
names(RNA.ref) <- c("chr", "start_pos", "end_pos", "strand", "ENSEMBL", "SYMBOL")
RNA.ref <- RNA.ref[!duplicated(RNA.ref$ENSEMBL),]
Y.genes <- RNA.ref[RNA.ref$chr == "Y", ]$ENSEMBL

##MAIN

#Getting Universe of data
universe <- unique(DEG.merged$ENTREZID)
#Filtering data
DEG.merged <- DEG.merged[(abs(DEG.merged$log2FoldChange) > L2FC.filter & !is.na(DEG.merged$log2FoldChange)) & 
                                    (DEG.merged$padj < p.adj.filter & !is.na(DEG.merged$padj)), ]
DEG.merged$Comparison <- gsub("\\.", "_", DEG.merged$Comparison)
DEG.merged <- DEG.merged[!DEG.merged$ENSEMBL %in% Y.genes, ]

### Doing enrichments

#REACTOME enrichment
DEG.Pathway <- compareCluster(ENTREZID~Direction+Comparison, data=DEG.merged, fun="enrichPathway", readable=T, universe = universe)
DEG.Pathway@compareClusterResult$minuslog10p.adjust <- -log10(DEG.Pathway@compareClusterResult$p.adjust)
DEG.Pathway.pw <- pairwise_termsim(DEG.Pathway)

#GO: Molecular function
DEG.mf <- compareCluster(ENTREZID~Direction+Comparison, data=DEG.merged, fun="enrichGO", OrgDb = org.Hs.eg.db, ont = "MF", readable=T, universe = universe) 
DEG.mf@compareClusterResult$minuslog10p.adjust <- -log10(DEG.mf@compareClusterResult$p.adjust)
DEG.mf.pw <- pairwise_termsim(DEG.mf)
DEG.mf.pw.simp <- simplify(DEG.mf.pw, cutoff=0.7, by="p.adjust", select_fun=min)

#GO: Biological process
DEG.bp <- compareCluster(ENTREZID~Direction+Comparison, data=DEG.merged, fun="enrichGO", OrgDb = org.Hs.eg.db, ont = "BP", readable=T, universe = universe) 
DEG.bp@compareClusterResult$minuslog10p.adjust <- -log10(DEG.bp@compareClusterResult$p.adjust)
DEG.bp.pw <- pairwise_termsim(DEG.bp)
DEG.bp.pw.simp <- simplify(DEG.bp.pw, cutoff=0.7, by="p.adjust", select_fun=min)

#GO: Celullar component
DEG.cc <- compareCluster(ENTREZID~Direction+Comparison, data=DEG.merged, fun="enrichGO", OrgDb = org.Hs.eg.db, ont = "CC", readable=T, universe = universe) 
DEG.cc@compareClusterResult$minuslog10p.adjust <- -log10(DEG.cc@compareClusterResult$p.adjust)
DEG.cc.pw <- pairwise_termsim(DEG.cc)
DEG.cc.pw.simp <- simplify(DEG.cc.pw, cutoff=0.7, by="p.adjust", select_fun=min)

#KEGG enrichment
DEG.kegg <- compareCluster(ENTREZID~Direction+Comparison, data=DEG.merged, fun="enrichKEGG", universe = universe) 
DEG.kegg@compareClusterResult$minuslog10p.adjust <- -log10(DEG.kegg@compareClusterResult$p.adjust)
DEG.kegg.pw <- pairwise_termsim(DEG.kegg)

#GSE enrichment
DEG.gse <- compareCluster(ENTREZID~Direction+Comparison, data=DEG.merged, fun=gse_fun) 
DEG.gse@compareClusterResult$minuslog10p.adjust <- -log10(DEG.gse@compareClusterResult$p.adjust)
DEG.gse.pw <- pairwise_termsim(DEG.gse)

##OUTPUT

#SAVING enrichments
write.csv(DEG.Pathway, 
          file="./output/tables/enrichments/DEG.Pathway.clusters.csv", 
          row.names = F, quote = F)
write.csv(DEG.mf, 
          file="./output/tables/enrichments/DEG.mf.clusters.csv", 
          row.names = F, quote = F)
write.csv(DEG.bp, 
          file="./output/tables/enrichments/DEG.bp.clusters.csv", 
          row.names = F, quote = F)
write.csv(DEG.cc, 
          file="./output/tables/enrichments/DEG.cc.clusters.csv", 
          row.names = F, quote = F)
write.csv(DEG.kegg, 
          file="./output/tables/enrichments/DEG.kegg.clusters.csv", 
          row.names = F, quote = F)
write.csv(DEG.gse, 
          file="./output/tables/enrichments/DEG.gse.clusters.csv", 
          row.names = F, quote = F)

#Saving enrichment objects
saveRDS(DEG.Pathway, "./output/objects/enrichments/DEG_clusters/Pathway.RDS")
saveRDS(DEG.mf, "./output/objects/enrichments/DEG_clusters/mf.RDS")
saveRDS(DEG.bp, "./output/objects/enrichments/DEG_clusters/bp.RDS")
saveRDS(DEG.cc, "./output/objects/enrichments/DEG_clusters/cc.RDS")
saveRDS(DEG.kegg, "./output/objects/enrichments/DEG_clusters/kegg.RDS")
saveRDS(DEG.gse, "./output/objects/enrichments/DEG_clusters/gse.RDS")

saveRDS(DEG.Pathway.pw, "./output/objects/enrichments/DEG_clusters/pw/Pathway.pw.RDS")
saveRDS(DEG.mf.pw.simp, "./output/objects/enrichments/DEG_clusters/pw/mf.pw.RDS")
saveRDS(DEG.bp.pw.simp, "./output/objects/enrichments/DEG_clusters/pw/bp.pw.RDS")
saveRDS(DEG.cc.pw.simp, "./output/objects/enrichments/DEG_clusters/pw/cc.pw.RDS")
saveRDS(DEG.kegg.pw, "./output/objects/enrichments/DEG_clusters/pw/kegg.pw.RDS")
saveRDS(DEG.gse.pw, "./output/objects/enrichments/DEG_clusters/pw/gse.pw.RDS")

write.csv(DEG.merged, 
          file="./output/tables/DEG/significant_cluster_comparisons_merged.csv", 
          row.names = F, quote = F)
write.table(DEG.Pathway, 
            file="./output/tables/enrichments/significant_cluster_comparisons_merged.txt", 
            row.names = F, quote = F, sep = "\t")
