==> 16_Formating_cell_proportions.R <==
#Script generated by: Aaron D. Ramirez-Sanchez 
#Goal: To change proportion names

##ENVIRONMENT

##FUNCTIONS

##INPUT

FACS.2groups <- read.table(file = "./data/raw/cell_proportions/scaled.FACS.2groupsVshared.txt")
FACS.6groups <- read.table(file = "./data/raw/cell_proportions/scaled.FACS.6groupsVshared.txt")

#Reading shared samples
shared.samples <- read.table(file = "./data/processed/metadata/shared.samplenames.new.txt")
shared.samples <- shared.samples$V1
shared.samples <- shared.samples[2:length(shared.samples)]

#Reading shared samples
oldshared.samples <- read.table(file = "./data/raw/metadata/shared.samplenames.old.txt")
oldshared.samples <- oldshared.samples$V1

##MAIN
#checking that sample names are ordered in the same way
identical(oldshared.samples, rownames(FACS.2groups))
identical(oldshared.samples, rownames(FACS.6groups))

rownames(FACS.2groups) <- shared.samples
rownames(FACS.6groups) <- shared.samples

##OUTPUT

write.table(FACS.2groups, 
            './output/tables/FACS/cell_proportions_scaled.2groups.txt', 
            sep='\t',quote=F,col.names=NA)
write.table(FACS.6groups, 
            './output/tables/FACS/cell_proportions_scaled.6groups.txt', 
            sep='\t',quote=F,col.names=NA)
