==> 21_Processing_decon-eQTL_tables.R <==
#Script generated by: Aaron D. Ramirez-Sanchez 
#Goal: To format metadata and create a sample_sample conversion

##ENVIRONMENT
library(dplyr)

##FUNCTIONS
processFile <- function(f) {
  df <- read.table(f, sep = "\t", header = T)
  return(df)
}
add.symbol <- function(df) {
  df <- left_join(df, gene.references, by = "ENSEMBL")
  return(df)
}
separating_eQTL <- function(df) {
  df$ENSEMBL <- gsub("_.*","",df$X)
  df$SNP <- gsub(".*_","",df$X)
  return(df)
}
add.FDR.for2 <- function(df) {
  df$FDR.Epithelial <- p.adjust(df$Epithelial_cells_pvalue, method = "fdr", n = length(df$Epithelial_cells_pvalue))
  df$FDR.Immune <- p.adjust(df$Immune_Cells_pvalue, method = "fdr", n = length(df$Immune_Cells_pvalue))
  return(df)
}
adding.sig.eQTL.for2 <- function(df) {
  df$sig.Epithelial <- df$Epithelial_cells_pvalue < 0.05
  df$sig.Immune <- df$Immune_Cells_pvalue < 0.05
  return(df)
}
plotting.pval.eQTL.for2 <- function(df) {
  p1 <- hist(df$Epithelial_cells_pvalue)
  p2 <- hist(df$Immune_Cells_pvalue)
  plots <- list(p1,p2)
  return(plots)
}

add.FDR.for6 <- function(df) {
  df$FDR.APC <- p.adjust(df$APC_P_pvalue, method = "fdr", n = length(df$APC_P_pvalue))
  df$FDR.TCRgd <- p.adjust(df$TCRgd_P_pvalue, method = "fdr", n = length(df$TCRgd_P_pvalue))
  df$FDR.CD4_CD8_dp <- p.adjust(df$CD4_CD8_dp_P_pvalue, method = "fdr", n = length(df$CD4_CD8_dp_P_pvalue))
  df$FDR.CD8 <- p.adjust(df$CD8_P_pvalue, method = "fdr", n = length(df$CD8_P_pvalue))
  df$FDR.CD4 <- p.adjust(df$CD4_P_pvalue, method = "fdr", n = length(df$CD4_P_pvalue))
  df$FDR.Small <- p.adjust(df$Small_CD45_P_pvalue, method = "fdr", n = length(df$Small_CD45_P_pvalue))
  df$FDR.Big <- p.adjust(df$Big_CD45_P_pvalue, method = "fdr", n = length(df$Big_CD45_P_pvalue))
  return(df)
}
adding.sig.eQTL.for6 <- function(df) {
  df$sig.APC <- df$APC_P_pvalue < 0.05
  df$sig.TCRgd <- df$TCRgd_P_pvalue < 0.05
  df$sig.CD4_CD8_dp <- df$CD4_CD8_dp_P_pvalue < 0.05
  df$sig.CD8 <- df$CD8_P_pvalue < 0.05
  df$sig.CD4 <- df$CD4_P_pvalue < 0.05
  df$sig.Small <- df$Small_CD45_P_pvalue < 0.05
  df$sig.Big <- df$Big_CD45_P_pvalue < 0.05
  return(df)
}
plotting.pval.eQTL.for6 <- function(df) {
  p1 <- hist(df$APC_P_pvalue) 
  p2 <- hist(df$TCRgd_P_pvalue) 
  p3 <- hist(df$CD4_CD8_dp_P_pvalue) 
  p4 <- hist(df$CD8_P_pvalue) 
  p5 <- hist(df$CD4_P_pvalue) 
  p6 <- hist(df$Small_CD45_P_pvalue) 
  p7 <- hist(df$Big_CD45_P_pvalue) 

  plots <- list(p1,p2,p3,p4,p5,p6,p7)
  return(plots)
}

##INPUT
#Reading enrichment objects
files <- dir("./data/processed/decon_eQTL", recursive=TRUE, full.names=TRUE, pattern="\\.csv$")
decon_eQTLs <- sapply(files, processFile)
names(decon_eQTLs) <- gsub("\\/decon.*", "", gsub(".*FACS", "FACS", files))

gene.references <- read.table(file = "/Users/R_projects/Oslo/results/eQTL/references/Homo_sapiens.GRCh37.75.genes.bed", sep = "\t")
names(gene.references) <- c("chr", "pos1", "pos2", "strand", "ENSEMBL", "SYMBOL")
gene.references <- gene.references[!duplicated(gene.references$ENSEMBL),]

##MAIN
filtered.decon_eQTLs <- sapply(decon_eQTLs, function(x){x[!duplicated(x$X),]})
filtered.decon_eQTLs <- sapply(filtered.decon_eQTLs, separating_eQTL)
filtered.decon_eQTLs <- sapply(filtered.decon_eQTLs, add.symbol)

decon_eQTLs_2 <- filtered.decon_eQTLs[1:4]
decon_eQTLs_6 <- filtered.decon_eQTLs[5:8]

decon_eQTLs_2 <- lapply(decon_eQTLs_2, add.FDR.for2)
decon_eQTLs_6 <- lapply(decon_eQTLs_6, add.FDR.for6)

decon_eQTLs_2 <- lapply(decon_eQTLs_2, adding.sig.eQTL.for2)
decon_eQTLs_6 <- lapply(decon_eQTLs_6, adding.sig.eQTL.for6)

plots_decon_eQTLs_2 <- lapply(decon_eQTLs_2, plotting.pval.eQTL.for2)
plots_decon_eQTLs_6 <- lapply(decon_eQTLs_6, plotting.pval.eQTL.for6)

table(decon_eQTLs_2$FACS2_RNAco$sig.Epithelial, decon_eQTLs_2$FACS2_RNAco$sig.Immune)
decon_eQTLs_2_table <- sapply(decon_eQTLs_2, function(df){table(df$sig.Epithelial, df$sig.Immune)})
row.names(decon_eQTLs_2_table) <- c("No","Epithelial","Immune","Both")

decon_eQTLs_2_sorted_imm <- lapply(decon_eQTLs_2, function(df){
  df <- df[order(df$Epithelial_cells_pvalue), ]
  df <- df[order(df$Immune_Cells_pvalue), ]
  df <- df[df$sig.Epithelial | df$sig.Immune, ]
  return(df)
  }
  )

decon_eQTLs_2_sorted_epi <- lapply(decon_eQTLs_2, function(df){
  df <- df[order(df$Immune_Cells_pvalue), ]
  df <- df[order(df$Epithelial_cells_pvalue), ]
  df <- df[df$sig.Epithelial | df$sig.Immune, ]
  return(df)
}
)

genes_imm <- lapply(decon_eQTLs_2, function(df){
  df <- df[order(df$Immune_Cells_pvalue), ]
  df <- df[df$sig.Immune == TRUE, ]
  df <- df[!duplicated(df$ENSEMBL), ]$SYMBOL
  return(df)
}
)

gene_epi <- lapply(decon_eQTLs_2, function(df){
  df <- df[order(df$Epithelial_cells_pvalue), ]
  df <- df[df$sig.Epithelial == TRUE, ]
  df <- df[!duplicated(df$ENSEMBL), ]$SYMBOL
  return(df)
}
)

##OUTPUT
pdf(file = "./output/plots/eQTLs_pvalues_histograms.pdf", width=7, height=7,  family = "ArialMT")
lapply(decon_eQTLs_2, plotting.pval.eQTL.for2)
lapply(decon_eQTLs_6, plotting.pval.eQTL.for6)
dev.off()

write.table(decon_eQTLs_2$FACS2_RNAco, 
            file="./output/tables/decon-eQTLs/decon_FACS2_RNA_condition.txt", 
            row.names = F, quote = F, sep = "\t")
write.table(decon_eQTLs_2$FACS2_RNAcr, 
            file="./output/tables/decon-eQTLs/decon_FACS2_RNA_crypts.txt", 
            row.names = F, quote = F, sep = "\t")
write.table(decon_eQTLs_2$FACS2_RNAm, 
            file="./output/tables/decon-eQTLs/decon_FACS2_RNA_marsh.txt", 
            row.names = F, quote = F, sep = "\t")
write.table(decon_eQTLs_2$FACS2_RNAmcc, 
            file="./output/tables/decon-eQTLs/decon_FACS2_RNA_mss.txt", 
            row.names = F, quote = F, sep = "\t")
