==> 23_PCA_plots_genotypes.R <==
#Script generated by: Aaron D. Ramirez-Sanchez 
#Goal: To plot genetic PCAs

##ENVIRONMENT

##FUNCTIONS

##INPUT
options(scipen=100, digits=3)

# read in the eigenvectors, produced in PLINK
eigenvec <- read.table('./data/processed/PCA/plink.eigenvec', header = FALSE, skip=0, sep = ' ')
rownames(eigenvec) <- eigenvec[,2]
eigenvec <- eigenvec[,3:ncol(eigenvec)]
colnames(eigenvec) <- paste('Principal Component ', c(1:20), sep = '')

# read in the PED data
PED <- read.table('./data/processed/PCA/20130606_g1k.ped', header = TRUE, skip = 0, sep = '\t')
PED <- PED[which(PED$Individual.ID %in% rownames(eigenvec)), ]

eigenvec$Population <- PED[match(rownames(eigenvec), PED$Individual.ID), ]$Population
eigenvec[is.na(eigenvec$Population),]$Population <- "Sample"

eigenvec$Population <- factor(eigenvec$Population, levels=c(
  "Sample",
  "ACB","ASW","ESN","GWD","LWK","MSL","YRI", #Africa 
  "CLM","MXL","PEL","PUR", #Latinoamerica
  "CDX","CHB","CHS","JPT","KHV", #East Asia
  "CEU","FIN","GBR","IBS","TSI", #European
  "BEB","GIH","ITU","PJL","STU" #South Asia
  ))
#PED.samples <- rownames(eigenvec[!rownames(eigenvec) %in% PED$Individual.ID,])
#PED.all <- data.frame(matrix(ncol = ncol(PED), nrow = nrow(eigenvec)))
#colnames(PED.all) <- colnames(PED)
#rownames(PED.all) <- rownames(eigenvec)

annotation.data <- read.delim(file = "./data/processed/metadata/extended_metadata.tsv", sep = "\t", row.names = 1)
eigenvec.complete <- eigenvec
eigenvec <- eigenvec[!(!(rownames(eigenvec) %in% annotation.data$sample) & eigenvec$Population == "Sample"),]
#PED <- PED[match(rownames(eigenvec), PED$Individual.ID),]
#all(PED$Individual.ID == rownames(eigenvec)) == TRUE
sample <- eigenvec[eigenvec$`Principal Component 1` < 0 & eigenvec$Population == "Sample",]

# set colours
require('RColorBrewer')

# from: http://www.internationalgenome.org/category/population/


col <- colorRampPalette(c(
  "#FC6700FF",
  "#315261","#315261","#315261","#315261","#315261","#315261","#315261",
  "#9260B5","#9260B5","#9260B5","#9260B5",
  "#8B8090","#8B8090","#8B8090","#8B8090","#8B8090",
  "royalblue","royalblue","royalblue","royalblue","royalblue",
  "#2A2B2DFF","#2A2B2DFF","#2A2B2DFF","#2A2B2DFF","#2A2B2DFF"))(length(unique(eigenvec$Population)))[factor(eigenvec$Population)]


##MAIN

##OUTPUT
# generate PCA bi-plots
eigenvec.short <- eigenvec[1:12]
project.pca <- eigenvec.short
summary(project.pca)


#par(mar = c(5,5,5,5), cex = 2.0,
#    cex.main = 7, cex.axis = 2.75, cex.lab = 2.75, mfrow = c(1,2))
par()
pdf(file = "./output/plots/PCA_genotypes.pdf", width=6, height=6,  family = "ArialMT")
plot(project.pca[,1], project.pca[,2],
     type = 'n',
     xlab = 'First component',
     ylab = 'Second component',
     font = 1,
     font.lab = 1)
points(project.pca[,1], project.pca[,2], col = col, pch = 20, cex = 1)
legend('bottomleft',
       c('Sample', 'Africans', 'Admixed Americans', 'East Asians',
         'Europeans', 'South Asians'),
       fill = c("#FC6700FF",
                "#315261",
                "#9260B5",
                "#8B8090",
                "royalblue",
                "#2A2B2DFF"))
dev.off()
